+++ R/helpers_ggcoefstats_graphics.R
@@ -65,12 +65,18 @@ ggcoefstats_label_maker <- function(tidy_df,
 
   # if the statistic is chi^2-value
   if (statistic == "c") {
+    # if not present, add NA column for dfs
+    if (!"df.error" %in% names(tidy_df)) tidy_df %<>% dplyr::mutate(df.error = Inf)
+
     tidy_df %<>%
       dplyr::mutate(
         label = paste0(
           "list(~widehat(italic(beta))==",
           specify_decimal_p(x = estimate, k = k),
-          ", ~italic(chi)^2==",
+          ", ~italic(chi)^2~",
+          "(",
+          specify_decimal_p(x = df.error, k = 0L),
+          ")==",
           specify_decimal_p(x = statistic, k = k),
           ", ~italic(p)=='",
           specify_decimal_p(x = p.value, k = k, p.value = TRUE),